pyproject.toml
[project]
name = "ccpd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gitpython>=3.1.44",
    "ipython>=9.2.0",
    "matplotlib>=3.2.2",
    "numpy>=1.18.5",
    "opencv-python>=4.1.1",
    "pandas>=1.1.4",
    "pillow>=7.1.2",
    "psutil>=7.0.0",
    "pyyaml>=5.3.1",
    "requests>=2.23.0",
    "scipy>=1.4.1",
    "seaborn>=0.11.0",
    "tensorboard>=2.4.1",
    "thop>=0.1.1",
    "torch>=1.7.0",
    "torchvision>=0.8.1",
    "tqdm>=4.64.0",
]
